Fixed logic error where current days' webcams would end up being half-processed

+++ webcams_process.sh
@@ -1,8 +1,6 @@
 #!/bin/bash
 logdir="/mnt/data/webcams/"
-
-#datestr=$(/bin/date -d yesterday +%Y/%m/%d)
-#datestr="2012-05-29"
+today=$(date +%Y/%m/%d)
 
 #Camera
 for A in $logdir*/
@@ -16,17 +14,20 @@ do
 			#Day
 			for D in ${C}*/
 			do
-				if find ${D}*.jpg -maxdepth 0 2>/dev/null | read
+				if [ "${D: -11:10}" != "$today" ]
 				then
-					vidstr=$(echo "${D}" | sed 's#/*$##')/log.avi
-					echo $vidstr
-					/usr/bin/mencoder "mf://${D}*.jpg" -mf fps=25 -o $vidstr -ovc lavc -lavcopts vcodec=mpeg4:vbitrate=700; 2>&1 > /dev/null
-					if [ -e $vidstr ]
+					if find ${D}*.jpg -maxdepth 0 2>/dev/null | read
 					then
-						rm ${D}$datestr/*.jpg;
-						#echo Remove images;
-					else
-						echo Error: Video $vidstr could not be verified.;
+						vidstr=$(echo "${D}" | sed 's#/*$##')/log.avi
+						echo $vidstr
+						/usr/bin/mencoder "mf://${D}*.jpg" -mf fps=25 -o $vidstr -ovc lavc -lavcopts vcodec=mpeg4:vbitrate=700; 2>&1 > /dev/null
+						if [ -e $vidstr ]
+						then
+							rm ${D}$datestr/*.jpg;
+							#echo Remove images;
+						else
+							echo Error: Video $vidstr could not be verified.;
+						fi
 					fi
 				fi
 			done
